Reorganized structure to make it easier to remove things once airflow gets them fixed
