+++ ec2/packer/airflow_python3/systemd/airflow-worker.service
@@ -1,0 +1,33 @@
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+[Unit]
+Description=Airflow webserver daemon
+After=network.target postgresql.service mysql.service redis.service rabbitmq-server.service
+Wants=postgresql.service mysql.service redis.service rabbitmq-server.service
+
+[Service]
+EnvironmentFile=/etc/profile.d/airflow.sh
+User=ubuntu
+Group=ubuntu
+Type=simple
+ExecStart=/bin/bash -c 'airflow worker --pid /run/airflow/worker.pid $* >> /home/ubuntu/airflow/logs/worker.log'
+Restart=on-failure
+RestartSec=5s
+PrivateTmp=true
+PIDFile=/run/airflow/worker.pid
+ExecReload=/bin/kill -s HUP $MAINPID
+ExecStop=/bin/kill -s TERM $MAINPID
+
+[Install]
+WantedBy=multi-user.target
